fix: full audit remediation — 33 issues across backend and frontend
Backend:
- #1 CRITICAL: SQL injection — parameterized all queries (instances, scanner, compress)
- #2: Atomic job counters (ProcessedImages, SavedBytes) with sync/atomic
- #3: (same as #2)
- #4: MediaItemIDs ownership validation before compression
- #5: Scanner concurrency control — per-instance mutex prevents concurrent scans
- #6: Backup race condition — unique .bak.{jobID} filenames
- #7: Scan timeout context (10 min) replaces context.Background()
- #8: deleteInstance now checks cascade errors
- #15: Scanner path deduplication
- #16: Clarifying comment on role mapping in mergeSettings
- #17: ON DELETE CASCADE migration for compression_results FK
- #18: Per-IP rate limiting middleware (30 req/s, burst 10)
- #19: walkMediaCoverDir error logging
- #26: Removed unused matched variable
- #27: Fixed missing columns in scanner INSERT
- #28: TODO comment for api_key encryption
- #29: Removed redundant WAL PRAGMA

Frontend:
- #2: fetchItems race condition — stale response discard via fetchId counter
- #9: Watcher race on compress config — settingsLoaded guard
- #10: Modal accessibility — role=dialog, aria-modal, focus trap, Escape close
- #11: Debounce timer cleanup on unmount
- #12: Preserve selection on page change
- #13: Tightened createInstance type (required fields, not Partial)
- #14: Filter params typed as '0'|'1' instead of number
- #20: Extracted ROLE_DEFAULTS etc. to shared constants/defaults.ts
- #21: console.warn in silent catch blocks
- #22: aria-labels on all select inputs
- #23: Comment explaining dirty guard in SettingsView deep watch
- #24: Added total_images and compressed_items to Stats type
- #25: Health endpoint already called by Layout.vue (verified)
- #31: Skip-to-main-content link
- #32: Generic setFilter type instead of any
- #33: int64 precision comment on min_size_kb fields

+++ internal/models/models.go
@@ -130,8 +130,8 @@ type CompressionJob struct {
 	Config         JobConfig  `json:"config"`
 	TotalItems     int        `json:"total_items"`
 	ProcessedItems int        `json:"processed_items"`
-	TotalImages    int        `json:"total_images"`
-	ProcessedImages int       `json:"processed_images"`
+	TotalImages    int64      `json:"total_images"`
+	ProcessedImages int64     `json:"processed_images"`
 	SavedBytes     int64      `json:"saved_bytes"`
 	ErrorCount     int        `json:"error_count"`
 	SkipCount      int        `json:"skip_count"`
